pyproject.toml
[project]
name = "game-python-opengl-platformer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "glfw>=2.10.0",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "pyglm>=2.8.3",
    "pyopengl>=3.1.10",
    "pyopengl-accelerate>=3.1.10",
]
